+++ lib/components/main_view.dart
@@ -8,6 +8,7 @@ Scaffold mainView(
   BuildContext context, {
   String? appBarTitle = 'changeMe',
   required Widget body,
+  bool showAppBar = true,
   Widget? appBarLeading,
   double? leadingWidth,
   bool automaticallyImplyLeading = true,
@@ -15,13 +16,15 @@ Scaffold mainView(
   PreferredSizeWidget? appbarBottom,
   Color? appBarBackgroundColor,
   Color? backgroundColor,
-  bool? extendBodyBehindAppBar = false,
+  bool extendBody = false,
+  bool extendBodyBehindAppBar = false,
   bool showFloatingActionButton = false,
   IconData? floatingActionButtonIcon = Icons.add,
   String? floatingActionButtonTooltip = "Create new",
   void Function()? onFloatingActionButtonPressed,
   FloatingActionButtonLocation floatingActionButtonLocation =
       FloatingActionButtonLocation.miniEndFloat,
+  FloatingActionButton? floatingActionButtonProps,
   Widget? bottomSheet,
   Widget? bottomNavigationBar,
   AlignmentDirectional persistentFooterAlignment =
@@ -32,21 +35,24 @@ Scaffold mainView(
     assert(onFloatingActionButtonPressed != null);
   }
   return Scaffold(
-    extendBodyBehindAppBar: extendBodyBehindAppBar!,
-    appBar: AppBar(
-      foregroundColor: Theme.of(context).appBarTheme.foregroundColor,
-      backgroundColor: appBarBackgroundColor ??
-          Theme.of(context).appBarTheme.backgroundColor,
-      title: Text(
-        appBarTitle!,
-        style: TextStyle(color: Theme.of(context).colorScheme.onSurface),
-      ),
-      automaticallyImplyLeading: automaticallyImplyLeading,
-      leading: appBarLeading,
-      leadingWidth: leadingWidth,
-      actions: appbarActions,
-      bottom: appbarBottom,
-    ),
+    extendBodyBehindAppBar: extendBodyBehindAppBar,
+    extendBody: extendBody,
+    appBar: showAppBar
+        ? AppBar(
+            foregroundColor: Theme.of(context).appBarTheme.foregroundColor,
+            backgroundColor: appBarBackgroundColor ??
+                Theme.of(context).appBarTheme.backgroundColor,
+            title: Text(
+              appBarTitle!,
+              style: TextStyle(color: Theme.of(context).colorScheme.onSurface),
+            ),
+            automaticallyImplyLeading: automaticallyImplyLeading,
+            leading: appBarLeading,
+            leadingWidth: leadingWidth,
+            actions: appbarActions,
+            bottom: appbarBottom,
+          )
+        : null,
     backgroundColor:
         backgroundColor ?? Theme.of(context).colorScheme.secondaryContainer,
     body: body,
@@ -56,6 +62,28 @@ Scaffold mainView(
     persistentFooterButtons: persistentFooterButtons,
     floatingActionButton: showFloatingActionButton == true
         ? FloatingActionButton(
+            key: floatingActionButtonProps?.key,
+            foregroundColor: floatingActionButtonProps?.foregroundColor,
+            backgroundColor: floatingActionButtonProps?.backgroundColor,
+            focusColor: floatingActionButtonProps?.focusColor,
+            hoverColor: floatingActionButtonProps?.hoverColor,
+            splashColor: floatingActionButtonProps?.splashColor,
+            heroTag: floatingActionButtonProps?.heroTag,
+            elevation: floatingActionButtonProps?.elevation,
+            focusElevation: floatingActionButtonProps?.focusElevation,
+            hoverElevation: floatingActionButtonProps?.hoverElevation,
+            highlightElevation: floatingActionButtonProps?.highlightElevation,
+            disabledElevation: floatingActionButtonProps?.disabledElevation,
+            mouseCursor: floatingActionButtonProps?.mouseCursor,
+            mini: floatingActionButtonProps?.mini ?? false,
+            shape: floatingActionButtonProps?.shape,
+            clipBehavior: floatingActionButtonProps?.clipBehavior ?? Clip.none,
+            focusNode: floatingActionButtonProps?.focusNode,
+            autofocus: floatingActionButtonProps?.autofocus ?? false,
+            materialTapTargetSize:
+                floatingActionButtonProps?.materialTapTargetSize,
+            isExtended: floatingActionButtonProps?.isExtended ?? false,
+            enableFeedback: true,
             tooltip: floatingActionButtonTooltip,
             onPressed: () => onFloatingActionButtonPressed!(),
             child: Icon(floatingActionButtonIcon),